function/requirements.txt
azure_storage==0.37.0
Flask==2.3.2
